pyproject.toml
[project]
name = "victron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "victron-ble>=0.9.2",
    "paho-mqtt>=2.1.0",
    "asyncio-mqtt>=0.16.2",
    "python-dotenv>=1.0.0",
]
